delete alert dialog added

+++ snatok/lib/views/manage_ad.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class ManageAd extends StatefulWidget{
+class ManageAd extends StatefulWidget {
   final Function _deleteAd;
   final List<Map<String, dynamic>> products;
 
@@ -10,40 +10,69 @@ class ManageAd extends StatefulWidget{
   State<StatefulWidget> createState() {
     return _ManageAdState();
   }
-
-  
 }
 
 class _ManageAdState extends State<ManageAd> {
-
   @override
   void initState() {
     super.initState();
   }
+
   @override
   Widget build(BuildContext context) {
     return ListView.builder(
       itemBuilder: (BuildContext context, int index) {
-        return ListTile(
-          leading: Icon(Icons.ac_unit),
-          title: Text(widget.products[index]['title']),
-          trailing: ButtonBar(
-            mainAxisSize: MainAxisSize.min,
-            children: <Widget>[
-              IconButton(
-                icon: Icon(Icons.edit),
-                onPressed: () {
-                  Navigator.of(context).pushNamed('/edit_ad/'+index.toString()) ;
-                },
+        return Column(
+          children: <Widget>[
+            ListTile(
+              leading: CircleAvatar(
+                backgroundImage: AssetImage(widget.products[index]['image']),
               ),
-              IconButton(
-                icon: Icon(Icons.delete),
-                onPressed: () {
-                  widget._deleteAd(index);
-                },
-              )
-            ],
-          ),
+              title: Text(widget.products[index]['title']),
+              subtitle: Text('\৳' + widget.products[index]['price'].toString()),
+              trailing: ButtonBar(
+                mainAxisSize: MainAxisSize.min,
+                children: <Widget>[
+                  IconButton(
+                    icon: Icon(Icons.edit),
+                    onPressed: () {
+                      Navigator.of(context)
+                          .pushNamed('/edit_ad/' + index.toString());
+                    },
+                  ),
+                  IconButton(
+                    icon: Icon(Icons.delete),
+                    onPressed: () {
+                      showDialog(
+                          builder: (BuildContext context) {
+                            return AlertDialog(
+                              title: Text('Are you sure to remove?'),
+                              content: Text('Once deleted can not be undone'),
+                              actions: <Widget>[
+                                FlatButton(
+                                  onPressed: () {
+                                    widget._deleteAd(index);
+                                    Navigator.pop(context);
+                                  },
+                                  child: Text('Yes'),
+                                ),
+                                FlatButton(
+                                  onPressed: () {
+                                    Navigator.pop(context);
+                                  },
+                                  child: Text('No'),
+                                ),
+                              ],
+                            );
+                          },
+                          context: context);
+                    },
+                  )
+                ],
+              ),
+            ),
+            Divider(),
+          ],
         );
       },
       itemCount: widget.products.length,
